Split up the requirements.txt file
We already started to do this by adding requirements-doc.txt and
requirements-test.txt.  This change continues on that path making
requirements-app.txt and then having requirements.txt include all
files.

This could be a good way to ensure that we only bring in the
requirements we need.  For example, for production app instances, we
probably just need to use requirements-app.txt.

This is just a quick start to the work.  I'm pretty sure some
requirements should be moved from one file to another.  Also we would
need to refactor our Dockerfile, maybe making one for app instances
hosted on amara.org that used requirements-app.txt and other for local
dev instances that used requirements.txt.